Cleanup headers
1.  This commit cleans up some of the headers so that they only
have the routines, declares, structs that are public.
2.  Revise the loopback functions to have vlp_ as the function name
prefixes for easier tracking in code.
